Create common component for 1st and 2nd passphrase generation
