+++ i18n/locales/en/common.json
@@ -88,6 +88,7 @@
   "Fee: {{amount}} LSK": "Fee: {{amount}} LSK",
   "Fee: {{fee}} LSK": "Fee: {{fee}} LSK",
   "Final confirmation": "Final confirmation",
+  "Get passphrase": "Get passphrase",
   "Get to your Dashboard": "Get to your Dashboard",
   "Go back to Dashboard": "Go back to Dashboard",
   "Go to \"More\" on your Sidebar to start the onboarding whenever you need.": "Go to \"More\" on your Sidebar to start the onboarding whenever you need.",
